Added testing for notification service
