+++ src/SFA.DAS.Commitments.Application/Services/ApprenticeshipFilterService.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 
 using SFA.DAS.Commitments.Api.Types.Apprenticeship;
