--- src/SFA.DAS.Commitments.Notification.WebJob.UnitTests/Program.cs
@@ -1,9 +1,0 @@
-﻿namespace SFA.DAS.Commitments.Notification.WebJob.UnitTests
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-        }
-    }
-}
