+++ src/SFA.DAS.Commitments.Notification.WebJob.UnitTests/packages.config
@@ -1,5 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="AutoFixture" version="3.50.1" targetFramework="net452" />
+  <package id="AutoFixture.NUnit3" version="3.50.1" targetFramework="net452" />
   <package id="Castle.Core" version="3.3.3" targetFramework="net452" />
   <package id="FluentAssertions" version="4.13.0" targetFramework="net452" />
   <package id="Microsoft.IdentityModel.Clients.ActiveDirectory" version="2.28.3" targetFramework="net452" />
@@ -10,5 +12,6 @@
   <package id="SFA.DAS.Account.Api.Client" version="1.0.0.30620" targetFramework="net452" />
   <package id="SFA.DAS.Account.Api.Types" version="1.0.0.30620" targetFramework="net452" />
   <package id="SFA.DAS.NLog.Logger" version="1.0.0.26805" targetFramework="net452" />
+  <package id="SFA.DAS.Notifications.Api.Client" version="1.1.0.12760" targetFramework="net452" />
   <package id="SFA.DAS.Notifications.Api.Types" version="1.1.0.12760" targetFramework="net452" />
 </packages>
