+++ src/SFA.DAS.Commitments.Notification.WebJob/Models/UserModel.cs
@@ -1,0 +1,14 @@
+using System.Collections.Generic;
+
+using SFA.DAS.EAS.Account.Api.Types;
+
+namespace SFA.DAS.Commitments.Notification.WebJob.Models
+{
+    internal class UserModel
+    {
+        public long AccountId { get; set; }
+
+        public IEnumerable<TeamMemberViewModel> Users { get; set; }
+
+    }
+}
